# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Workout Log 공개 API",
    "version": "1.0.0",
    "description": "본인이 발급한 개인 액세스 토큰(PAT)으로 자기 데이터에 접근하는 API다.\n\n**기본은 거부다.** 여기 적힌 경로만 PAT로 닿을 수 있고, 그 밖의 모든 경로는\n401을 준다. 설정 변경·계정 관리·데이터 삭제·가져오기는 어느 스코프로도 열리지\n않는다.\n\n**응답 스키마는 담지 않는다.** 22개 엔드포인트의 응답 형태를 손으로 유지하면\n곧 거짓말이 된다. 이 문서가 답하는 것은 무엇을 어떤 권한으로 부를 수 있고,\n어떤 한도와 계약이 걸리는가다."
  },
  "servers": [
    {
      "url": "https://workout-log-eight.vercel.app",
      "description": "프로덕션"
    }
  ],
  "components": {
    "securitySchemes": {
      "personalAccessToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "`Authorization: Bearer wlpat_…` 형식. 설정 > 계정에서 발급한다.\n\n평문은 발급 화면에서 **한 번만** 보인다 — 서버는 SHA-256 해시만 저장하므로\n잃어버리면 폐기하고 재발급해야 한다. 폐기는 즉시 적용된다(캐시 없음).\n\n스코프는 `read`(기본)와 `read_write` 둘이다.\n\n⚠️ 브라우저에서 세션 쿠키와 함께 보내면 **쿠키가 이긴다**. PAT 클라이언트는\n쿠키를 보내지 않으므로 문제되지 않는다."
      }
    }
  },
  "security": [
    {
      "personalAccessToken": []
    }
  ],
...
  "paths": {
...
    "/api/stats/bundle": {
      "get": {
        "summary": "통계 묶음",
        "description": "홈·통계 화면이 한 번에 쓰는 집계 묶음.",
        "security": [
          {
            "personalAccessToken": []
          }
        ],
        "x-required-scopes": [
          "read",
          "read_write"
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "401": {
            "description": "토큰이 없거나, 폐기됐거나, 이 경로가 공개 표면 밖이다(기본은 거부)."
          },
          "403": {
            "description": "토큰 스코프가 이 작업을 허용하지 않는다(read 토큰의 쓰기)."
          },
          "429": {
            "description": "요청 한도 초과. 읽기 120/분, 쓰기 30/분. `Retry-After` 헤더를 볼 것."
          }
        }
      }
    },
    "/api/stats/e1rm": {
      "get": {
        "summary": "e1RM 추이",
        "description": "운동별 추정 1RM 시계열. 웜업 세트는 제외된다.",
        "security": [
          {
            "personalAccessToken": []
          }
        ],
        "x-required-scopes": [
          "read",
          "read_write"
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "401": {
            "description": "토큰이 없거나, 폐기됐거나, 이 경로가 공개 표면 밖이다(기본은 거부)."
          },
          "403": {
            "description": "토큰 스코프가 이 작업을 허용하지 않는다(read 토큰의 쓰기)."
          },
          "429": {
            "description": "요청 한도 초과. 읽기 120/분, 쓰기 30/분. `Retry-After` 헤더를 볼 것."
          }
        }
      }
    },
    "/api/stats/prs": {
      "get": {
        "summary": "개인 최고 기록",
        "description": "종목별 PR과 향상도.",
        "security": [
          {
            "personalAccessToken": []
          }
        ],
        "x-required-scopes": [
          "read",
          "read_write"
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "401": {
            "description": "토큰이 없거나, 폐기됐거나, 이 경로가 공개 표면 밖이다(기본은 거부)."
          },
          "403": {
            "description": "토큰 스코프가 이 작업을 허용하지 않는다(read 토큰의 쓰기)."
          },
          "429": {
            "description": "요청 한도 초과. 읽기 120/분, 쓰기 30/분. `Retry-After` 헤더를 볼 것."
          }
        }
      }
    },
    "/api/stats/strength-summary": {
      "get": {
        "summary": "근력 요약",
        "description": "3대 토탈과 체중 대비 배율. 배율은 **그 시점 체중** 기준이다.",
        "security": [
          {
            "personalAccessToken": []
          }
        ],
        "x-required-scopes": [
          "read",
          "read_write"
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "401": {
            "description": "토큰이 없거나, 폐기됐거나, 이 경로가 공개 표면 밖이다(기본은 거부)."
          },
          "403": {
            "description": "토큰 스코프가 이 작업을 허용하지 않는다(read 토큰의 쓰기)."
          },
          "429": {
            "description": "요청 한도 초과. 읽기 120/분, 쓰기 30/분. `Retry-After` 헤더를 볼 것."
          }
        }
      }
    },
    "/api/stats/volume": {
      "get": {
        "summary": "볼륨 집계",
        "description": "기간 볼륨 총계.",
        "security": [
          {
            "personalAccessToken": []
          }
        ],
        "x-required-scopes": [
          "read",
          "read_write"
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "401": {
            "description": "토큰이 없거나, 폐기됐거나, 이 경로가 공개 표면 밖이다(기본은 거부)."
          },
          "403": {
            "description": "토큰 스코프가 이 작업을 허용하지 않는다(read 토큰의 쓰기)."
          },
          "429": {
            "description": "요청 한도 초과. 읽기 120/분, 쓰기 30/분. `Retry-After` 헤더를 볼 것."
          }
        }
      }
    },
    "/api/stats/volume-series": {
      "get": {
        "summary": "볼륨 시계열",
        "description": "주 단위 볼륨 추이.",
        "security": [
          {
            "personalAccessToken": []
          }
        ],
        "x-required-scopes": [
          "read",
          "read_write"
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "401": {
            "description": "토큰이 없거나, 폐기됐거나, 이 경로가 공개 표면 밖이다(기본은 거부)."
          },
          "403": {
            "description": "토큰 스코프가 이 작업을 허용하지 않는다(read 토큰의 쓰기)."
          },
          "429": {
            "description": "요청 한도 초과. 읽기 120/분, 쓰기 30/분. `Retry-After` 헤더를 볼 것."
          }
        }
      }
    },
    "/api/stats/muscle-freshness": {
      "get": {
        "summary": "부위별 신선도",
        "description": "결정론 감쇠 모델로 추정한 부위별 회복도. `groups[].contributions`에 그 값을 만든 세션이 함께 온다 — 점수만 주고 근거를 숨기지 않는다. `capacityKg === 0`은 **회복 완료가 아니라 기록 없음**이다.",
        "security": [
          {
            "personalAccessToken": []
          }
        ],
        "x-required-scopes": [
          "read",
          "read_write"
        ],
        "responses": {
          "200": {
            "description": "성공"
          },
          "401": {
            "description": "토큰이 없거나, 폐기됐거나, 이 경로가 공개 표면 밖이다(기본은 거부)."
          },
          "403": {
            "description": "토큰 스코프가 이 작업을 허용하지 않는다(read 토큰의 쓰기)."
          },
          "429": {
            "description": "요청 한도 초과. 읽기 120/분, 쓰기 30/분. `Retry-After` 헤더를 볼 것."
          }
        }
      }
    },
...
  }
}